ByteBufferBackedInputStream を ZipHelper 以外から利用できないように修正する 併せて、 DataChannel メッセージを文字列に変換する際に、 CharsetDecoder を利用するように修正する
